src/use-cases/get-user-profile.ts
import { UsersRepository } from '@/repositories/users-repository'
import { User } from '@prisma/client'
import { ResourceNotFoundError } from './errors/resource-not-found'

interface GetUserProfileUseCaseReq {
  user_id: string
}

interface GetUserProfileUseCaseRes {
  user: User
}

export class GetUserProfileUseCase {
  constructor(private usersRepository: UsersRepository) {}

  async execute({
    user_id,
  }: GetUserProfileUseCaseReq): Promise<GetUserProfileUseCaseRes> {
    const user = await this.usersRepository.findById(user_id)

    if (!user) {
      throw new ResourceNotFoundError()
    }

    return { user }
  }
}
